Yamux update

+++ libp2p-muxer-yamux/src/main/kotlin/org/erwinkok/libp2p/muxer/yamux/YamuxStreamId.kt
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 Erwin Kok. BSD-3-Clause license. See LICENSE file for more details.
 package org.erwinkok.libp2p.muxer.yamux
 
-data class YamuxStreamId(val initiator: Boolean, val id: Long) {
+data class YamuxStreamId(val initiator: Boolean, val id: Int) {
     override fun toString(): String {
         return String.format("stream%08x/%s", id, if (initiator) "initiator" else "responder")
     }
